feat: Adjust sector change logic for suppliers
The logic for determining if a supplier can change a sector for an attendee has been refined. Previously, it checked if any allowed sector was *not* included in the attendee's sectors.

Now, it checks if the number of allowed sectors for the supplier is greater than the number of sectors currently assigned to the attendee. This provides a more flexible approach to sector management.